Rename webhook.git.kube.ci -> webhooks.git.kube.ci
