--- apis/webhook/register.go
@@ -1,3 +1,0 @@
-package webhook
-
-const GroupName = "webhook.git.kube.ci"
