+++ apis/webhooks/register.go
@@ -1,0 +1,3 @@
+package webhooks
+
+const GroupName = "webhooks.git.kube.ci"
