+++ apis/webhooks/v1alpha1/doc.go
@@ -1,5 +1,5 @@
 // Package v1alpha1 is the v1alpha1 version of the API.
 
 // +k8s:deepcopy-gen=package
-// +groupName=webhook.git.kube.ci
+// +groupName=webhooks.git.kube.ci
 package v1alpha1
