+++ apis/webhooks/v1alpha1/register.go
@@ -6,7 +6,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
-const GroupName = "webhook.git.kube.ci"
+const GroupName = "webhooks.git.kube.ci"
 
 var SchemeGroupVersion = schema.GroupVersion{Group: GroupName, Version: "v1alpha1"}
