+++ chart/git-apiserver/templates/apiregistration.yaml
@@ -26,14 +26,14 @@ spec:
 apiVersion: apiregistration.k8s.io/v1beta1
 kind: APIService
 metadata:
-  name: v1alpha1.webhook.git.kube.ci
+  name: v1alpha1.webhooks.git.kube.ci
   labels:
     chart: "{{ .Chart.Name }}-{{ .Chart.Version }}"
     app: "{{ template "git-apiserver.name" . }}"
     heritage: "{{ .Release.Service }}"
     release: "{{ .Release.Name }}"
 spec:
-  group: webhook.git.kube.ci
+  group: webhooks.git.kube.ci
   version: v1alpha1
   service:
     namespace: {{ .Release.Namespace }}
