+++ client/clientset/versioned/typed/webhook/v1alpha1/webhook_client.go
@@ -19,7 +19,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	v1alpha1 "github.com/kube-ci/git-apiserver/apis/webhook/v1alpha1"
+	v1alpha1 "github.com/kube-ci/git-apiserver/apis/webhooks/v1alpha1"
 	"github.com/kube-ci/git-apiserver/client/clientset/versioned/scheme"
 	serializer "k8s.io/apimachinery/pkg/runtime/serializer"
 	rest "k8s.io/client-go/rest"
@@ -29,7 +29,7 @@ type WebhookV1alpha1Interface interface {
 	RESTClient() rest.Interface
 }
 
-// WebhookV1alpha1Client is used to interact with features provided by the webhook.git.kube.ci group.
+// WebhookV1alpha1Client is used to interact with features provided by the webhooks.git.kube.ci group.
 type WebhookV1alpha1Client struct {
 	restClient rest.Interface
 }
