+++ hack/deploy/apiservices.yaml
@@ -19,12 +19,12 @@ spec:
 apiVersion: apiregistration.k8s.io/v1beta1
 kind: APIService
 metadata:
-  name: v1alpha1.webhook.git.kube.ci
+  name: v1alpha1.webhooks.git.kube.ci
   labels:
     app: git-apiserver
 spec:
   caBundle: ${SERVICE_SERVING_CERT_CA}
-  group: webhook.git.kube.ci
+  group: webhooks.git.kube.ci
   groupPriorityMinimum: 1000
   versionPriority: 15
   service:
