+++ hack/deploy/rbac-list.yaml
@@ -130,14 +130,14 @@ subjects:
   name: ${GIT_APISERVER_SERVICE_ACCOUNT}
   namespace: ${GIT_APISERVER_NAMESPACE}
 ---
-# Give system:anonymous user access to apis/webhook.git.kube.ci/v1alpha1/githubevents path
+# Give system:anonymous user access to apis/webhooks.git.kube.ci/v1alpha1/githubevents path
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
   name: git-apiserver-webhook
 rules:
 - apiGroups:
-  - webhook.git.kube.ci
+  - webhooks.git.kube.ci
   resources:
   - githubevents
   verbs: ["create"]
