+++ hack/dev/apiregistration.yaml
@@ -53,12 +53,12 @@ spec:
 apiVersion: apiregistration.k8s.io/v1beta1
 kind: APIService
 metadata:
-  name: v1alpha1.webhook.git.kube.ci
+  name: v1alpha1.webhooks.git.kube.ci
   labels:
     app: git-dev-apiserver
 spec:
   insecureSkipTLSVerify: true
-  group: webhook.git.kube.ci
+  group: webhooks.git.kube.ci
   groupPriorityMinimum: 1000
   versionPriority: 15
   service:
